removed; fixed column bg color, summary getValue
